requirements.txt
# Database
mysql-connector-python==8.2.0
PyMySQL==1.1.0

# Security and encryption
bcrypt==4.1.2
cryptography==42.0.8

# Utilities
python-dotenv==1.0.0
colorama==0.4.6

# Development
black==23.11.0
flake8==6.1.0